+++ src/app/pages/LiveRadioPage/index.stories.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter } from 'react-router-dom';
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withServicesKnob } from '@bbc/psammead-storybook-helpers';
-import { RadioPage } from '..';
+import { LiveRadioPage } from '..';
 import indonesia from '#data/indonesia/bbc_indonesian_radio/liveradio.json';
 import korean from '#data/korean/bbc_korean_radio/liveradio.json';
 import tigrinya from '#data/tigrinya/bbc_tigrinya_radio/liveradio.json';
@@ -45,7 +45,7 @@ storiesOf('Pages|Radio Page', module)
   )
   .add('default', ({ service }) => (
     <BrowserRouter>
-      <RadioPage
+      <LiveRadioPage
         match={matchFixtures(service)}
         pageData={liveRadioFixtures[service]}
         status={status}
