+++ src/app/pages/index.js
@@ -4,7 +4,7 @@ import _ErrorPage from './ErrorPage';
 import _FrontPage from './FrontPage';
 import _MediaAssetPage from './MediaAssetPage';
 import _PhotoGalleryPage from './PhotoGalleryPage';
-import _RadioPage from './RadioPage';
+import _LiveRadioPage from './LiveRadioPage';
 import _OnDemandRadioPage from './OnDemandRadioPage';
 import _StoryPage from './StoryPage';
 import withContexts from '#containers/PageHandlers/withContexts';
@@ -31,6 +31,6 @@ export const ErrorPage = applyErrorPageHandlers(_ErrorPage);
 export const FrontPage = pipe(applyBasicPageHandlers, withVariant)(_FrontPage);
 export const MediaAssetPage = applyBasicPageHandlers(_MediaAssetPage);
 export const PhotoGalleryPage = applyBasicPageHandlers(_PhotoGalleryPage);
-export const RadioPage = applyBasicPageHandlers(_RadioPage);
+export const LiveRadioPage = applyBasicPageHandlers(_LiveRadioPage);
 export const OnDemandRadioPage = applyBasicPageHandlers(_OnDemandRadioPage);
 export const StoryPage = applyBasicPageHandlers(_StoryPage);
