+++ src/app/routes/index.js
@@ -1,7 +1,7 @@
 import article from './article';
 import cpsAsset from './cpsAsset';
 import home from './home';
-import liveRadio from './radio';
+import liveRadio from './liveRadio';
 import onDemandRadio from './onDemandRadio';
 import error from './error';
 import errorNoRouteMatch from './errorNoRouteMatch';
