+++ src/app/routes/liveRadio/getInitialData/addIdsToBlocks.js
