+++ src/app/routes/liveRadio/index.js
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { RadioPage } from '#pages';
+import { LiveRadioPage } from '#pages';
 import pageIsSame from '../utils/pageIsSame';
 import { liveRadioPath } from '../utils/regex';
 import getInitialData from './getInitialData';
@@ -7,7 +7,7 @@ import getInitialData from './getInitialData';
 export default {
   path: liveRadioPath,
   exact: true,
-  component: memo(RadioPage, pageIsSame),
+  component: memo(LiveRadioPage, pageIsSame),
   getInitialData,
   pageType: 'media',
 };
